apps/sidecar/pyproject.toml
[project]
name = "sidecar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "joshuabrandes", email = "47532061+joshuabrandes@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "uvicorn>=0.42.0",
    "marker-pdf>=1.10.2",
    "marker>=2.1.3",
    "fitz>=0.0.1.dev2",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]

[project.scripts]
# sidecar = "sidecar:main"
sidecar = "sidecar.main:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
